mitosheet: move optimize to pro, add missing init
